Refactor factories
